+++ reductstore/src/storage/query.rs
@@ -17,7 +17,7 @@ use reduct_base::error::ReductError;
 use reduct_base::unprocessable_entity;
 use std::sync::{Arc, Mutex, RwLock};
 use std::thread::sleep;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 use tokio::sync::mpsc::Receiver;
 
 pub(crate) type QueryRx = Receiver<Result<RecordReader, ReductError>>;
@@ -51,7 +51,7 @@ pub(super) fn spawn_query_task(
 ) -> (QueryRx, TaskHandle<()>) {
     let (tx, rx) = tokio::sync::mpsc::channel(QUERY_BUFFER_SIZE);
 
-    // we spawn a new task to run the query outside of hierarchical task group to avoid deadlocks
+    // we spawn a new task to run the query outside hierarchical task group to avoid deadlocks
     let query = Arc::new(Mutex::new(query));
     let handle = shared_isolated("spawn_query_task", "spawn query task", move || {
         trace!("Query task for '{}' running", task_group);
@@ -62,6 +62,8 @@ pub(super) fn spawn_query_task(
             let block_manager = block_manager.clone();
             let tx = tx.clone();
 
+            // the task return None if the loop must be stopped
+            // we do it for each iteration so we don't need to take the whole worker for a long query
             let task = shared(&group.clone(), "query iteration", move || {
                 if tx.is_closed() {
                     trace!("Query '{}' task channel closed", group);
@@ -70,8 +72,7 @@ pub(super) fn spawn_query_task(
 
                 if tx.capacity() == 0 {
                     trace!("Query '{}' task channel full", group);
-                    sleep(Duration::from_micros(10));
-                    return Some(());
+                    return Some(Duration::from_millis(10));
                 }
 
                 let next_result = query.lock().unwrap().next(block_manager.clone());
@@ -88,19 +89,18 @@ pub(super) fn spawn_query_task(
                     if err.status == NoContent && options.continuous {
                         // continuous query will never be done
                         // but we don't want to flood the channel and wait for the receiver
-                        let now = Instant::now();
-                        while tx.capacity() < tx.max_capacity() && now.elapsed() < options.ttl {
-                            sleep(Duration::from_millis(10));
-                        }
-                    } else {
-                        trace!("Query task done for '{}'", group);
-                        return None;
+                        return Some(options.ttl / 4);
                     }
+
+                    trace!("Query task done for '{}'", group);
+                    return None;
                 }
-                Some(())
+                Some(Duration::from_millis(0))
             });
 
-            if task.wait().is_none() {
+            if let Some(sleep_duration) = task.wait() {
+                sleep(sleep_duration);
+            } else {
                 break;
             }
         }
@@ -111,7 +111,6 @@ pub(super) fn spawn_query_task(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::core::logger::Logger;
     use crate::storage::block_manager::block_index::BlockIndex;
     use crate::storage::proto::Record;
     use prost_wkt_types::Timestamp;
@@ -176,13 +175,12 @@ mod tests {
         assert_eq!(rx.recv().await.unwrap().unwrap().timestamp(), 0);
         assert_eq!(rx.recv().await.unwrap().unwrap().timestamp(), 1);
         assert_eq!(rx.recv().await.unwrap().err().unwrap().status, NoContent);
-        assert!(timeout(Duration::from_millis(1000), handle).await.is_ok());
+        assert_eq!(timeout(Duration::from_millis(1000), handle).await, Ok(()));
     }
 
     #[rstest]
     #[tokio::test]
     async fn test_query_task_continuous_ok(block_manager: Arc<RwLock<BlockManager>>) {
-        Logger::init("TRACE");
         let options = QueryOptions {
             ttl: Duration::from_millis(50),
             continuous: true,